code styling again
It's frustrating to need to deal with code styling while prototyping but that's the way it is.  I'm not satisfied with the code styling set up. It needs automation both in the IDE and outside and it needs consistency.  Applying fixes to Vue components is not working in Webstorm.  We also want the JS Standard styling but that is difficult to get to work because the build system from the Vue CLI setup has put Prettier in place.  Parking this for now.