WIP - prior to another refactor
The following explores a problem about how to manage data between the competing Vue requirements.

The app has 30+ “pages” that are identical except for a page identifier.  Each “page” is has a Vue route and each route refers to the page’s component. The component files are generated from configuration.  Configuration rarely changes and the generation is started by the developer whenever a change is needed.
The page components also set up of lifecycle hooks to link in the helper and navigation guards that prevent users from leaving a form in the middle of edits.

A page may have a form containing data values from the database. For example demographics.firstName.
The form is composed of child components: form elements. This is where the values are displayed and edited.
The form is disabled until a button is clicked that begins the edit. 
On load the values in the form elements need to show what is in the database or the default value.
The default value may be defined in the configuration. If not then the default value is based on the data type.
To allow for editing these values need to be instantiated, in the form element’s data, when the page loads so that Vue can bind to them and be reactive.
The form element watches for changes to the input and broadcasts an event on each change.
The helper tracks the changes and is ready to save the data.
There are two possible paths once the edit is started. Either the edit is canceled or saved. In either case, there is a call to the server to get the latest data. This data is retrieved.

At this point I need to push the current data down into the page/form/form elements.  It is the last that is important.  The form element gets its data through the initial value process described elsewhere here.
The need to update the data suggests the form element will have a second channel to get its value.  Two channels is complicated and ugly.  There should be a single source of truth for form elements when they are not editing.  Unfortunately, if the form element uses the same DOM input element to display and edit the value then it also uses the same v-model. This needs to be bound to a local data property.  

The helper needs to push the value into the form element on changes. Yet also, the form element needs to pull this value whenever it needs it.

 I have considered using the Vuex data storage but these values can not be tied to a v-model because the v-model needs to have full write access to the model. Vuex data storage values can not be modified by anyone except the storage itself. Others need to make a commit request to make changes.  Thus, using Vuex is not a clear winner.

I’ve been trying to manage this data in some form of shared object. Shared between the helper class and the components.  But this shared object is not reactive.

Maybe a global ‘data has changed’ event could be broadcast and the form elements could respond by asking the helper for their latest value.  They could then push the update into their local storage.  But be careful because the form element watches for changes to the input value and broadcasts an event when this happens. Don’t want to broadcast if we’re updating after a database refresh.

The plan is to have a getter for a data value in the helper. The form element will use this to get the initial value. The form element will use this when told to refresh.